address few comments
