+++ hadoop-ozone/ozonefs-common/src/main/java/org/apache/hadoop/fs/ozone/OzoneClientUtils.java
@@ -25,7 +25,6 @@ import org.apache.hadoop.hdds.conf.ConfigurationSource;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
 import org.apache.hadoop.fs.FileChecksum;
 import org.apache.hadoop.hdds.scm.OzoneClientConfig;
-import org.apache.hadoop.net.NetUtils;
 import org.apache.hadoop.ozone.client.checksum.BaseFileChecksumHelper;
 import org.apache.hadoop.ozone.client.ObjectStore;
 import org.apache.hadoop.ozone.client.OzoneBucket;
@@ -40,8 +39,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.net.UnknownHostException;
 import java.util.Objects;
 import java.util.Set;
 
@@ -253,14 +250,4 @@ public final class OzoneClientUtils {
     }
     return limitVal;
   }
-
-  public static void resolveOmHost(String omHost, int omPort) {
-    InetSocketAddress omHostAddress = NetUtils.createSocketAddr(omHost, omPort);
-    if (omHostAddress.isUnresolved()) {
-      throw new IllegalArgumentException(
-          "Cannot resolve OM host " + omHost + " in the URI",
-          new UnknownHostException());
-    }
-  }
-
 }
